weather-app/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
google-cloud-bigquery==3.12.0
google-cloud-storage==2.11.0
cloud-sql-python-connector[pg8000]
SQLAlchemy==2.0.21
google-cloud-error-reporting